move SlotLockProvider errors, misc cleanup
